+++ src/main/java/net/kaczmarzyk/spring/data/jpa/web/RepeatedJoinFetchResolver.java
@@ -16,43 +16,32 @@
 package net.kaczmarzyk.spring.data.jpa.web;
 
 import net.kaczmarzyk.spring.data.jpa.domain.Conjunction;
-import net.kaczmarzyk.spring.data.jpa.web.annotation.Join;
 import net.kaczmarzyk.spring.data.jpa.web.annotation.JoinFetch;
-import net.kaczmarzyk.spring.data.jpa.web.annotation.Joins;
+import net.kaczmarzyk.spring.data.jpa.web.annotation.RepeatedJoinFetch;
 import org.springframework.data.jpa.domain.Specification;
 
 import java.lang.annotation.Annotation;
 import java.util.ArrayList;
 import java.util.Collection;
 
-
 /**
- * @deprecated
- * This is going to be removed with {@link net.kaczmarzyk.spring.data.jpa.web.annotation.Joins}
- *
- * @author Tomasz Kaczmarzyk
+ * @author Jakub Radlica
  */
-@Deprecated
-class JoinsSpecificationResolver implements SpecificationResolver<Joins> {
+public class RepeatedJoinFetchResolver implements SpecificationResolver<RepeatedJoinFetch> {
 
 	private JoinFetchSpecificationResolver joinFetchSpecificationResolver = new JoinFetchSpecificationResolver();
-	private JoinSpecificationResolver joinSpecificationResolver = new JoinSpecificationResolver();
 
 	@Override
 	public Class<? extends Annotation> getSupportedSpecificationDefinition() {
-		return Joins.class;
+		return RepeatedJoinFetch.class;
 	}
 
 	@Override
-	public Specification<Object> buildSpecification(WebRequestProcessingContext context, Joins joinsDef) {
+	public Specification<Object> buildSpecification(WebRequestProcessingContext context, RepeatedJoinFetch def) {
 		Collection<Specification<Object>> joins = new ArrayList<>();
 
-		for (JoinFetch fetchDef : joinsDef.fetch()) {
+		for (JoinFetch fetchDef : def.value()) {
 			joins.add(joinFetchSpecificationResolver.buildSpecification(context, fetchDef));
-		}
-
-		for (Join joinDef : joinsDef.value()) {
-			joins.add(joinSpecificationResolver.buildSpecification(context, joinDef));
 		}
 
 		return new Conjunction<>(joins);