+++ src/main/java/net/kaczmarzyk/spring/data/jpa/web/annotation/Join.java
@@ -15,12 +15,8 @@
  */
 package net.kaczmarzyk.spring.data.jpa.web.annotation;
 
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
 import javax.persistence.criteria.JoinType;
+import java.lang.annotation.*;
 
 /**
  * Specifies a join part of a query, e.g. {@code select c from Customer c inner join c.addresses a}
@@ -29,6 +25,7 @@ import javax.persistence.criteria.JoinType;
  */
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ ElementType.PARAMETER, ElementType.TYPE })
+@Repeatable(RepeatedJoin.class)
 public @interface Join {
 
 	/**