+++ src/test/java/net/kaczmarzyk/spring/data/jpa/web/SpecificationArgumentResolverTest.java
@@ -53,26 +53,22 @@ public class SpecificationArgumentResolverTest extends ResolverTestBase {
             .contains(new net.kaczmarzyk.spring.data.jpa.domain.JoinFetch<Object>(new String[] { "fetch1", "fetch2" }, JoinType.LEFT, true));
     }
     
-//    @Test
-//    public void resolvesRepeatedFetchJoins() throws Exception {
-//    	MethodParameter param = MethodParameter.forExecutable(testMethod("testMethod_repeatedFetch"), 0);
-//        NativeWebRequest req = mock(NativeWebRequest.class);
-//        when(req.getParameterValues("path1")).thenReturn(new String[] { "value1" });
-//
-//        Specification<?> resolved = (Specification<?>) resolver.resolveArgument(param, null, req, null);
-//
-//        assertThat(resolved)
-//            .isInstanceOf(Conjunction.class);
-//        
-//        Collection<Specification<?>> innerSpecs = ReflectionUtils.get(resolved, "innerSpecs");
-//        
-//        assertThat(innerSpecs)
-//            .hasSize(2)
-//            .contains(new Like<Object>("path1", new String[] { "value1" }))
-//            .contains(new Conjunction<Object>(
-//            		new net.kaczmarzyk.spring.data.jpa.domain.JoinFetch<Object>(new String[] { "fetch1" }, JoinType.LEFT),
-//            		new net.kaczmarzyk.spring.data.jpa.domain.JoinFetch<Object>(new String[] { "fetch2" }, JoinType.INNER)));
-//    }
+    @Test
+    public void resolvesRepeatedFetchJoins() throws Exception {
+    	MethodParameter param = MethodParameter.forExecutable(testMethod("testMethod_repeatedFetch"), 0);
+        NativeWebRequest req = mock(NativeWebRequest.class);
+        QueryContext queryCtx = new WebRequestQueryContext(req);
+        when(req.getParameterValues("path1")).thenReturn(new String[] { "value1" });
+
+        Specification<?> resolved = (Specification<?>) resolver.resolveArgument(param, null, req, null);
+
+        assertThat(innerSpecs(resolved))
+            .hasSize(2)
+            .contains(new Like<>(queryCtx, "path1", new String[]{ "value1" }))
+            .contains(new Conjunction<>(
+                    new net.kaczmarzyk.spring.data.jpa.domain.JoinFetch<>(new String[]{ "fetch1" }, JoinType.LEFT, true),
+                    new net.kaczmarzyk.spring.data.jpa.domain.JoinFetch<>(new String[]{ "fetch2" }, JoinType.INNER, true)));
+    }
     
     @Test
     public void resolvesJoinContainerWithJoinFetch() throws Exception {
@@ -142,25 +138,33 @@ public class SpecificationArgumentResolverTest extends ResolverTestBase {
     public void resolvesJoinFetchForAnnotatedInterface() throws Exception {
         MethodParameter param = MethodParameter.forExecutable(testMethod("testMethodWithCustomSpec", CustomSpec.class), 0);
         NativeWebRequest req = mock(NativeWebRequest.class);
+        QueryContext queryCtx = new WebRequestQueryContext(req);
         when(req.getParameterValues("path1")).thenReturn(new String[] { "value1" });
 
         Specification<?> resolved = (Specification<?>) resolver.resolveArgument(param, null, req, null);
 
-        assertThat(resolved)
-            .isInstanceOf(CustomSpec.class); // TODO better assertions
+        assertThat(innerSpecs(resolved))
+                .hasSize(2)
+                .contains(new Like<Object>(queryCtx, "path1", new String[] { "value1" }))
+                .contains(new net.kaczmarzyk.spring.data.jpa.domain.JoinFetch<Object>(new String[] { "fetch1", "fetch2" }, JoinType.LEFT, true));
     }
     
-//    @Test
-//    public void resolvesRepeatedJoinFetchForAnnotatedInterface() throws Exception {
-//        MethodParameter param = MethodParameter.forExecutable(testMethod("testMethodWithCustomSpec_repeatedFetch", CustomSpecRepeatedFetch.class), 0);
-//        NativeWebRequest req = mock(NativeWebRequest.class);
-//        when(req.getParameterValues("path1")).thenReturn(new String[] { "value1" });
-//
-//        Specification<?> resolved = (Specification<?>) resolver.resolveArgument(param, null, req, null);
-//
-//        assertThat(resolved)
-//            .isInstanceOf(CustomSpecRepeatedFetch.class); // TODO better assertions
-//    }
+    @Test
+    public void resolvesRepeatedJoinFetchForAnnotatedInterface() throws Exception {
+        MethodParameter param = MethodParameter.forExecutable(testMethod("testMethodWithCustomSpec_repeatedFetch", CustomSpecRepeatedFetch.class), 0);
+        NativeWebRequest req = mock(NativeWebRequest.class);
+        QueryContext queryCtx = new WebRequestQueryContext(req);
+        when(req.getParameterValues("path1")).thenReturn(new String[] { "value1" });
+
+        Specification<?> resolved = (Specification<?>) resolver.resolveArgument(param, null, req, null);
+
+        assertThat(innerSpecs(resolved))
+                .hasSize(2)
+                .contains(new Like<>(queryCtx, "path1", new String[]{ "value1" }))
+                .contains(new Conjunction<>(
+                        new net.kaczmarzyk.spring.data.jpa.domain.JoinFetch<>(new String[]{ "fetch1" }, JoinType.LEFT, true),
+                        new net.kaczmarzyk.spring.data.jpa.domain.JoinFetch<>(new String[]{ "fetch2" }, JoinType.LEFT, true)));
+    }
     
     @Test
     public void resolvesJoinContainerForAnnotatedInterface() throws Exception {
@@ -184,11 +188,11 @@ public class SpecificationArgumentResolverTest extends ResolverTestBase {
     public static interface CustomSpec extends Specification<Object> {
     }
     
-//    @JoinFetch(paths = { "fetch1" })
-//    @JoinFetch(paths = { "fetch2" })
-//    @Spec(path = "path1", spec = Like.class)
-//    public static interface CustomSpecRepeatedFetch extends Specification<Object> {
-//    }
+    @JoinFetch(paths = { "fetch1" })
+    @JoinFetch(paths = { "fetch2" })
+    @Spec(path = "path1", spec = Like.class)
+    public static interface CustomSpecRepeatedFetch extends Specification<Object> {
+    }
     
     @Joins(fetch = {
     	@JoinFetch(paths = { "fetch1" }),
@@ -203,8 +207,8 @@ public class SpecificationArgumentResolverTest extends ResolverTestBase {
     	public void testMethodWithCustomSpec(CustomSpec spec) {
         }
     	
-//    	public void testMethodWithCustomSpec_repeatedFetch(CustomSpecRepeatedFetch spec) {
-//        }
+    	public void testMethodWithCustomSpec_repeatedFetch(CustomSpecRepeatedFetch spec) {
+        }
     	
     	public void testMethodWithCustomSpec_joinContainer(CustomSpecJoinContainer spec) {
         }
@@ -214,11 +218,11 @@ public class SpecificationArgumentResolverTest extends ResolverTestBase {
                 @Spec(path = "path1", spec = Like.class) Specification<Object> spec) {
         }
         
-//        public void testMethod_repeatedFetch(
-//                @JoinFetch(paths = { "fetch1" })
-//                @JoinFetch(paths = { "fetch2" }, joinType = JoinType.INNER)
-//                @Spec(path = "path1", spec = Like.class) Specification<Object> spec) {
-//        }
+        public void testMethod_repeatedFetch(
+                @JoinFetch(paths = { "fetch1" })
+                @JoinFetch(paths = { "fetch2" }, joinType = JoinType.INNER)
+                @Spec(path = "path1", spec = Like.class) Specification<Object> spec) {
+        }
         
         public void testMethod_joinContainerWithJoinFetch(
                 @Joins(fetch = {